refactor: 삭제 기록 관리하는 DeleteHistories 생성에 따른 리팩토링

+++ src/main/java/qna/service/QnAService.java
@@ -35,6 +35,7 @@ public class QnAService {
     @Transactional
     public void deleteQuestion(User loginUser, long questionId) throws CannotDeleteException {
         Question question = findQuestionById(questionId);
+        question.validateOwner(loginUser);
 
         DeleteHistories deleteHistories = question.deleteBy(loginUser);
